Make FieldGroupWrapper a component
