+++ lib/bali/form_builder_helpers/password_fields.rb
@@ -4,7 +4,7 @@ module Bali
   module FormBuilderHelpers
     module PasswordFields
       def password_field_group(method, options = {})
-        FieldGroupWrapper.render @template, self, method, options do
+        @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
           password_field(method, options)
         end
       end
