+++ lib/bali/form_builder_helpers/percentage_fields.rb
@@ -11,7 +11,7 @@ module Bali
           pattern_type: :number_with_commas
         )
 
-        FieldGroupWrapper.render @template, self, method, options do
+        @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
           text_field(method, options)
         end
       end
