+++ lib/bali/form_builder_helpers/slim_select_fields.rb
@@ -4,7 +4,7 @@ module Bali
   module FormBuilderHelpers
     module SlimSelectFields
       def slim_select_group(method, values, options = {}, html_options = {})
-        FieldGroupWrapper.render @template, self, method, options do
+        @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
           slim_select_field(method, values, options, html_options)
         end
       end
